Removed link for kaomoji.ru/en
Removed as they added crypto mining for revenue generation. To give credit where credit is due they are transparent about it and state they have done it AND they did add the controls to stop the process, user notification of it is way at the bottom of the page and not at the top where it should be.